third_party/WebKit/LayoutTests/inspector/coverage/multiple-instances-merge.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script src="coverage-test.js"></script>
<script src="resources/coverage.js"></script>

<link rel="stylesheet" type="text/css" href="resources/highlight-in-source.css">

<script>
async function test()
{
    CoverageTestRunner.startCoverage();
    await TestRunner.evaluateInPagePromise("performActions(); frames[0].performActionsInFrame()");
    await CoverageTestRunner.stopCoverage();

    var coverageView = self.runtime.sharedInstance(Coverage.CoverageView);
    var dataGrid = coverageView._listView._dataGrid;
    for (var child of dataGrid.rootNode().children) {
        var data = child._coverageInfo;
        var url = TestRunner.formatters.formatAsURL(data.url());
        if (url.endsWith("-test.js") || url.endsWith(".html"))
            continue;
        var type = Coverage.CoverageListView._typeToString(data.type());
        TestRunner.addResult(`${url} ${type} used: ${data.usedSize()} unused: ${data.unusedSize()} total: ${data.size()}`);
    }

    await CoverageTestRunner.dumpDecorations("highlight-in-source.css")
    await CoverageTestRunner.dumpDecorations("coverage.js")

    TestRunner.completeTest();
}

</script>
</head>

<body onload="runTest()">
<iframe src="resources/subframe.html"></iframe>
<p class="class">
Tests the coverage list view after finishing recording in the Coverage view.
</p>

</body>
</html>
